Removed unnecessary check.
Originally htmlvalidator was meant to speed up the validation process of a school project. The check that has been removed (footer must contain an author tag, identical to the meta author one) was not an error per se, but a violation of the project's guidelines. The reason behind this removal is that the aforementioned guidelines most likely won't apply to any user outside my project's group, and the project is now over.